# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "GoWVP Owl 的 HTTP API 文档，包含设备、通道、GB28181、PTZ、录像与系统接口。\n\n鉴权方式：\n大多数管理接口需要在请求头中携带 `Authorization: Bearer \u003ctoken\u003e`。\n可先调用 `/login/key` 获取 RSA 公钥，再调用 `/login` 完成登录。\n\n返回约定：\n1. 成功时通常返回业务对象，或 `{ \"msg\": \"ok\" }`\n2. 失败时通常返回 `{ \"msg\": \"错误原因\" }`\n3. GB28181 查询类接口通常会同时返回结构化结果与原始 XML",
        "title": "GoWVP Owl API",
        "contact": {},
        "version": "0.0.1"
    },
    "basePath": "/",
    "paths": {
...
        "/webhook/on_play": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ZLMWebhook"
                ],
                "summary": "ZLM 播放鉴权回调",
                "parameters": [
                    {
                        "description": "播放信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.onPublishInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.DefaultOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/webhook/on_publish": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ZLMWebhook"
                ],
                "summary": "ZLM 推流鉴权回调",
                "parameters": [
                    {
                        "description": "推流信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.onPublishInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.onPublishOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/webhook/on_record_mp4": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ZLMWebhook"
                ],
                "summary": "ZLM MP4 录像完成回调",
                "parameters": [
                    {
                        "description": "录像完成信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.onRecordMP4Input"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.DefaultOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/webhook/on_rtp_server_timeout": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ZLMWebhook"
                ],
                "summary": "ZLM RTP 服务器超时回调",
                "parameters": [
                    {
                        "description": "RTP 超时信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.onRTPServerTimeoutInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.DefaultOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/webhook/on_server_started": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ZLMWebhook"
                ],
                "summary": "ZLM 服务启动回调",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.DefaultOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/webhook/on_stream_changed": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ZLMWebhook"
                ],
                "summary": "ZLM 流注册变更回调",
                "parameters": [
                    {
                        "description": "流变更信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.onStreamChangedInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.DefaultOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/webhook/on_stream_none_reader": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ZLMWebhook"
                ],
                "summary": "ZLM 无人观看回调",
                "parameters": [
                    {
                        "description": "流信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.onStreamNoneReaderInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.onStreamNoneReaderOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/webhook/on_stream_not_found": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ZLMWebhook"
                ],
                "summary": "ZLM 流不存在回调",
                "parameters": [
                    {
                        "description": "流不存在信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.onStreamNotFoundInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.DefaultOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "api.DefaultOutput": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "错误代码，0 代表允许",
                    "type": "integer",
                    "example": 0
                },
                "msg": {
                    "description": "结果说明",
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "api.SwaggerErrorResponse": {
            "type": "object",
            "properties": {
                "msg": {
                    "description": "错误信息",
                    "type": "string",
                    "example": "参数错误"
                }
            }
        },
...
        "api.onPublishInput": {
            "type": "object",
            "properties": {
                "app": {
                    "description": "应用名",
                    "type": "string",
                    "example": "live"
                },
                "id": {
                    "description": "TCP 链接唯一 ID",
                    "type": "string",
                    "example": "1"
                },
                "ip": {
                    "description": "推流或播放端 IP",
                    "type": "string",
                    "example": "192.168.1.10"
                },
                "mediaServerId": {
                    "description": "流媒体服务器 ID",
                    "type": "string",
                    "example": "local"
                },
                "params": {
                    "description": "URL 参数",
                    "type": "string",
                    "example": "token=abc"
                },
                "port": {
                    "description": "端口号",
                    "type": "integer",
                    "example": 50000
                },
                "schema": {
                    "description": "协议，rtsp/rtmp/http 等",
                    "type": "string",
                    "example": "rtsp"
                },
                "stream": {
                    "description": "流 ID",
                    "type": "string",
                    "example": "camera001"
                },
                "vhost": {
                    "description": "虚拟主机",
                    "type": "string",
                    "example": "__defaultVhost__"
                }
            }
        },
        "api.onPublishOutput": {
            "type": "object",
            "properties": {
                "add_mute_audio": {
                    "type": "boolean"
                },
                "auto_close": {
                    "type": "boolean"
                },
                "code": {
                    "description": "错误代码，0 代表允许",
                    "type": "integer",
                    "example": 0
                },
                "continue_push_ms": {
                    "type": "integer"
                },
                "enable_audio": {
                    "type": "boolean"
                },
                "enable_fmp4": {
                    "type": "boolean"
                },
                "enable_hls": {
                    "type": "boolean"
                },
                "enable_hls_fmp4": {
                    "type": "boolean"
                },
                "enable_mp4": {
                    "type": "boolean"
                },
                "enable_rtmp": {
                    "type": "boolean"
                },
                "enable_rtsp": {
                    "type": "boolean"
                },
                "enable_ts": {
                    "type": "boolean"
                },
                "hls_save_path": {
                    "type": "string"
                },
                "modify_stamp": {
                    "type": "boolean"
                },
                "mp4_as_player": {
                    "type": "boolean"
                },
                "mp4_max_second": {
                    "type": "integer"
                },
                "mp4_save_path": {
                    "type": "string"
                },
                "msg": {
                    "description": "结果说明",
                    "type": "string",
                    "example": "success"
                },
                "stream_replace": {
                    "type": "string"
                }
            }
        },
        "api.onRTPServerTimeoutInput": {
            "type": "object",
            "properties": {
                "local_port": {
                    "description": "RTP 本地端口",
                    "type": "integer",
                    "example": 30000
                },
                "mediaServerId": {
                    "description": "流媒体服务器 ID",
                    "type": "string",
                    "example": "local"
                },
                "re_use_port": {
                    "description": "是否复用端口",
                    "type": "boolean",
                    "example": false
                },
                "ssrc": {
                    "description": "SSRC",
                    "type": "integer",
                    "example": 12345678
                },
                "stream_id": {
                    "description": "流 ID",
                    "type": "string",
                    "example": "camera001"
                },
                "tcp_mode": {
                    "description": "TCP 模式",
                    "type": "integer",
                    "example": 0
                }
            }
        },
        "api.onRecordMP4Input": {
            "type": "object",
            "properties": {
                "app": {
                    "description": "录制流应用名",
                    "type": "string",
                    "example": "live"
                },
                "file_name": {
                    "description": "文件名",
                    "type": "string",
                    "example": "camera001_20260320.mp4"
                },
                "file_path": {
                    "description": "文件绝对路径",
                    "type": "string",
                    "example": "/data/record/camera001.mp4"
                },
                "file_size": {
                    "description": "文件大小，字节",
                    "type": "integer",
                    "example": 1024000
                },
                "folder": {
                    "description": "文件目录",
                    "type": "string",
                    "example": "/data/record"
                },
                "mediaServerId": {
                    "description": "流媒体服务器 ID",
                    "type": "string",
                    "example": "local"
                },
                "start_time": {
                    "description": "开始录制时间戳（秒）",
                    "type": "integer",
                    "example": 1710931200
                },
                "stream": {
                    "description": "录制流 ID",
                    "type": "string",
                    "example": "camera001"
                },
                "time_len": {
                    "description": "录制时长，秒",
                    "type": "number",
                    "example": 60
                },
                "url": {
                    "description": "点播相对 URL",
                    "type": "string",
                    "example": "/record/live/camera001.mp4"
                },
                "vhost": {
                    "description": "流虚拟主机",
                    "type": "string",
                    "example": "__defaultVhost__"
                }
            }
        },
...
        "api.onStreamChangedInput": {
            "type": "object",
            "properties": {
                "aliveSecond": {
                    "description": "存活时长，秒",
                    "type": "integer",
                    "example": 0
                },
                "app": {
                    "description": "应用名",
                    "type": "string",
                    "example": "live"
                },
                "app_name": {
                    "description": "以下字段为 lalmax 新增",
                    "type": "string",
                    "example": "live"
                },
                "bytesSpeed": {
                    "description": "当前字节速度",
                    "type": "integer",
                    "example": 0
                },
                "createStamp": {
                    "description": "创建时间戳",
                    "type": "integer",
                    "example": 1710931200
                },
                "mediaServerId": {
                    "description": "流媒体服务器 ID",
                    "type": "string",
                    "example": "local"
                },
                "originSock": {
                    "$ref": "#/definitions/api.OriginSock"
                },
                "originType": {
                    "description": "源类型",
                    "type": "integer",
                    "example": 2
                },
                "originTypeStr": {
                    "description": "源类型文本",
                    "type": "string",
                    "example": "rtsp_push"
                },
                "originUrl": {
                    "description": "源地址",
                    "type": "string",
                    "example": ""
                },
                "readerCount": {
                    "description": "当前协议观看人数",
                    "type": "integer",
                    "example": 0
                },
                "regist": {
                    "description": "true 表示注册，false 表示注销",
                    "type": "boolean",
                    "example": true
                },
                "schema": {
                    "description": "协议",
                    "type": "string",
                    "example": "rtsp"
                },
                "stream": {
                    "description": "流 ID",
                    "type": "string",
                    "example": "test"
                },
                "stream_name": {
                    "description": "流名称",
                    "type": "string",
                    "example": "test"
                },
                "totalReaderCount": {
                    "description": "总观看人数",
                    "type": "integer",
                    "example": 0
                },
                "tracks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.Tracks"
                    }
                },
                "vhost": {
                    "description": "虚拟主机",
                    "type": "string",
                    "example": "__defaultVhost__"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}